CORE-1337 fix conflicts
