+++ src/Spryker/Client/RabbitMq/RabbitMqConfig.php
@@ -6,6 +6,8 @@
 
 namespace Spryker\Client\RabbitMq;
 
+use Generated\Shared\Transfer\QueueConnectionTransfer;
+use Generated\Shared\Transfer\RabbitMqOptionTransfer;
 use Spryker\Client\Kernel\AbstractBundleConfig;
 use Spryker\Shared\RabbitMq\RabbitMqConstants;
 
@@ -13,9 +15,39 @@ class RabbitMqConfig extends AbstractBundleConfig
 {
 
     /**
+     * @return \Generated\Shared\Transfer\QueueConnectionTransfer
+     */
+    public function getQueueConnection()
+    {
+        $queueConfig = $this->getQueueConnectionConfig();
+
+        $connectionTransfer = new QueueConnectionTransfer();
+        $connectionTransfer->setHost($queueConfig['host']);
+        $connectionTransfer->setPort($queueConfig['port']);
+        $connectionTransfer->setUsername($queueConfig['username']);
+        $connectionTransfer->setPassword($queueConfig['password']);
+        $connectionTransfer->setVirtualHost($queueConfig['virtualHost']);
+
+        $connectionTransfer->setQueueOptionCollection($this->getQueueOptions());
+
+        return $connectionTransfer;
+    }
+
+    /**
+     * @return \ArrayObject
+     */
+    protected function getQueueOptions()
+    {
+        $queueOptionCollection = new \ArrayObject();
+        $queueOptionCollection->append(new RabbitMqOptionTransfer());
+
+        return $queueOptionCollection;
+    }
+
+    /**
      * @return array
      */
-    public function getQueueConnectionConfig()
+    protected function getQueueConnectionConfig()
     {
         return [
             'host' => $this->get(RabbitMqConstants::RABBITMQ_HOST),
