+++ src/Spryker/Client/RabbitMq/RabbitMqDependencyProvider.php
@@ -17,51 +17,4 @@ use Spryker\Shared\RabbitMq\RabbitMqConstants;
 
 class RabbitMqDependencyProvider extends AbstractDependencyProvider
 {
-
-    const QUEUE_CONNECTION = 'queue connection config';
-
-    /**
-     * @param \Spryker\Client\Kernel\Container $container
-     *
-     * @return \Spryker\Client\Kernel\Container
-     */
-    public function provideServiceLayerDependencies(Container $container)
-    {
-        $container[static::QUEUE_CONNECTION] = function () {
-            return $this->getQueueConnection();
-        };
-
-        return $container;
-    }
-
-    /**
-     * @return \Generated\Shared\Transfer\QueueConnectionTransfer
-     */
-    public function getQueueConnection()
-    {
-        $queueConfig = $this->getConfig()->getQueueConnectionConfig();
-
-        $connectionTransfer = new QueueConnectionTransfer();
-        $connectionTransfer->setHost($queueConfig['host']);
-        $connectionTransfer->setPort($queueConfig['port']);
-        $connectionTransfer->setUsername($queueConfig['username']);
-        $connectionTransfer->setPassword($queueConfig['password']);
-        $connectionTransfer->setVirtualHost($queueConfig['virtualHost']);
-
-        $connectionTransfer->setQueueOptionCollection($this->getQueueOptions());
-
-        return $connectionTransfer;
-    }
-
-    /**
-     * @return \ArrayObject
-     */
-    protected function getQueueOptions()
-    {
-        $queueOptionCollection = new ArrayObject();
-        $queueOptionCollection->append(new RabbitMqOptionTransfer());
-
-        return $queueOptionCollection;
-    }
-
 }
